handle running on uninstalled revisions and selecting no cycles
git-svn-id: http://dev.oxf.freelogy.org/svn/polony/polony-tools/trunk@70 1b2bb62c-e52f-0410-8f12-f050cf38ac25

+++ www/installrevision.sh
@@ -9,7 +9,13 @@ echo "Installing revision $1 from $2."
 export dir=/usr/local/polony-tools/"$1"
 export svnrepos="$2"
 
-srun -N$SLURM_NNODES sh -c '
+if srun -N$SLURM_NNODES ls "$dir"/.tested
+then
+  echo Revision "$1" already installed in "$dir"/.
+  exit 0
+fi
+
+srun -N$SLURM_NNODES --job-name="r$1" --output="$dir".log sh -c '
 rm -rf "$dir" && \
 mkdir "$dir" && \
 cd "$dir" && \
